pyproject.toml
[project]
name = "reporting"
version = "0.1.0"
description = "Simple markdown/html report templating with loops, tables, lists and media."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "reporting maintainers" }]
keywords = ["report", "template", "markdown", "html"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
]
dependencies = [
  "rich",
  "pillow",
  "matplotlib",
]

[project.scripts]
reporting = "reporting.cli:main"

[project.urls]
Homepage = "https://example.com/reporting"
Repository = "https://example.com/reporting"
Documentation = "https://example.com/reporting/docs"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
